+++ flow/core/src/main/java/br/net/mirante/singular/flow/core/ProcessInstance.java
@@ -261,7 +261,7 @@ public abstract class ProcessInstance {
         return getInternalEntity().getEndDate();
     }
 
-    public final Integer getEntityCod() {
+    public final Long getEntityCod() {
         return entity.getCod();
     }
